# compaction: add support for concurrent manual compactions