--- a/compaction_picker_test.go
+++ b/compaction_picker_test.go
@@ -304,8 +304,8 @@ func TestCompactionPickerTargetLevel(t *testing.T) {
 				manual := &manualCompaction{
 					done:  make(chan error, 1),
 					level: startLevel,
-					start: iStart,
-					end:   iEnd,
+					start: iStart.UserKey,
+					end:   iEnd.UserKey,
 				}
 
 				pc, retryLater := pickerByScore.pickManual(compactionEnv{