--- a/format_major_version_test.go
+++ b/format_major_version_test.go
@@ -68,7 +68,7 @@ func testBasicDB(d *DB) error {
 	if err := d.Flush(); err != nil {
 		return err
 	}
-	if err := d.Compact(nil, []byte("\xff")); err != nil {
+	if err := d.Compact(nil, []byte("\xff"), false); err != nil {
 		return err
 	}