--- a/internal/datadriven/datadriven.go
+++ b/internal/datadriven/datadriven.go
@@ -25,6 +25,8 @@ import (
 	"strconv"
 	"strings"
 	"testing"
+
+	"github.com/cockroachdb/errors"
 )
 
 var (
@@ -229,47 +231,28 @@ type TestData struct {
 // td.ScanArgs(t, "arg3", &i2, &i3, &i4)
 func (td *TestData) ScanArgs(t *testing.T, key string, dests ...interface{}) {
 	t.Helper()
-	var arg CmdArg
-	for i := range td.CmdArgs {
-		if td.CmdArgs[i].Key == key {
-			arg = td.CmdArgs[i]
-			break
-		}
-	}
-	if arg.Key == "" {
+	arg := td.findArg(key)
+	if arg == nil {
 		t.Fatalf("missing argument: %s", key)
 	}
-	if len(dests) != len(arg.Vals) {
-		t.Fatalf("%s: got %d destinations, but %d values", arg.Key, len(dests), len(arg.Vals))
+	err := arg.scan(dests...)
+	if err != nil {
+		t.Fatal(err)
 	}
+}
 
-	for i := range dests {
-		val := arg.Vals[i]
-		switch dest := dests[i].(type) {
-		case *string:
-			*dest = val
-		case *int:
-			n, err := strconv.ParseInt(val, 10, 64)
-			if err != nil {
-				t.Fatal(err)
-			}
-			*dest = int(n) // assume 64bit ints
-		case *uint64:
-			n, err := strconv.ParseUint(val, 10, 64)
-			if err != nil {
-				t.Fatal(err)
-			}
-			*dest = n
-		case *bool:
-			b, err := strconv.ParseBool(val)
-			if err != nil {
-				t.Fatal(err)
-			}
-			*dest = b
-		default:
-			t.Fatalf("unsupported type %T for destination #%d (might be easy to add it)", dest, i+1)
+// HasArg determines if `key` appears in CmdArgs.
+func (td *TestData) HasArg(key string) bool {
+	return td.findArg(key) != nil
+}
+
+func (td *TestData) findArg(key string) *CmdArg {
+	for i := range td.CmdArgs {
+		if td.CmdArgs[i].Key == key {
+			return &td.CmdArgs[i]
 		}
 	}
+	return nil
 }
 
 // CmdArg contains information about an argument on the directive line. An
@@ -293,6 +276,41 @@ func (arg CmdArg) String() string {
 	default:
 		return fmt.Sprintf("%s=(%s)", arg.Key, strings.Join(arg.Vals, ", "))
 	}
+}
+
+func (arg CmdArg) scan(dests ...interface{}) error {
+	if len(dests) != len(arg.Vals) {
+		return errors.Errorf("%s: got %d destinations, but %d values", arg.Key, len(dests), len(arg.Vals))
+	}
+
+	for i := range dests {
+		val := arg.Vals[i]
+		switch dest := dests[i].(type) {
+		case *string:
+			*dest = val
+		case *int:
+			n, err := strconv.ParseInt(val, 10, 64)
+			if err != nil {
+				return err
+			}
+			*dest = int(n) // assume 64bit ints
+		case *uint64:
+			n, err := strconv.ParseUint(val, 10, 64)
+			if err != nil {
+				return err
+			}
+			*dest = n
+		case *bool:
+			b, err := strconv.ParseBool(val)
+			if err != nil {
+				return err
+			}
+			*dest = b
+		default:
+			return errors.Errorf("unsupported type %T for destination #%d (might be easy to add it)", dest, i+1)
+		}
+	}
+	return nil
 }
 
 // Fatalf wraps a fatal testing error with test file position information, so