This PR introduces concurrent execution for manual compactions - previously done one at a time, level by level.

This is achieved by splitting a single manual compaction for a level into
multiple compactions with non-overlapping key ranges. When compacted, the
output ranges of these compactions do not conflict and can thus be
performed in parallel.

A new parameter has been added to db.Compact, which determines whether or
not to split a single manual compaction into its encompassing non-overlapping
keyranges. Moreover, the changes in this PR rely on automatic compactions
being disabled when compactions are split to avoid conflicts.

Resolves https://github.com/cockroachdb/pebble/issues/1404