@mufeez-amjad 2022-02-16
I benchmarked my changes against master on a ~150GB LSM, here are the results:

### Summary

Roughly a 30% decrease in time taken to execute a manual compaction.

| Branch      | Time | 
| :---        |    :----   | 
| `master`      | real    550m32.809s<br>user    589m52.219ss<br>sys     36m25.205s   | 
|  `concurrent-manual` |  real    384m56.152s<br>user    373m0.258s<br>sys     28m25.139s  | 

<details>
  <summary>LSM</summary>

**Test data:**
```
__level_____count____size___score______in__ingest(sz_cnt)____move(sz_cnt)___write(sz_cnt)____read___r-amp___w-amp
    WAL         3   138 M       -     0 B       -       -       -       -   138 M       -       -       -     0.0
      0     69428    84 G    3.02     0 B     0 B       0     0 B       0     0 B       0     0 B       0     0.0
      1         0     0 B    0.00     0 B     0 B       0     0 B       0     0 B       0     0 B       0     0.0
      2      2478   9.7 G    5.45     0 B     0 B       0     0 B       0     0 B       0     0 B       0     0.0
      3      1975    12 G    5.72     0 B     0 B       0     0 B       0     0 B       0     0 B       0     0.0
      4      1497    15 G    5.73     0 B     0 B       0     0 B       0     0 B       0     0 B       0     0.0
      5       961    18 G    0.87     0 B     0 B       0     0 B       0     0 B       0     0 B       0     0.0
      6       703    15 G       -     0 B     0 B       0     0 B       0     0 B       0     0 B       0     0.0
  total     77042   154 G       -   138 M     0 B       0     0 B       0   138 M       0     0 B       0     1.0
  flush         0
compact         0   583 G     0 B       0          (size == estimated-debt, score = in-progress-bytes, in = num-in-progress)
  ctype         0       0       0       0       0  (default, delete, elision, move, read)
 memtbl        39   144 M
zmemtbl         0     0 B
   ztbl         0     0 B
 bcache         0     0 B    0.0%  (score == hit-rate)
 tcache         0     0 B    0.0%  (score == hit-rate)
  snaps         0       -       0  (score == earliest seq num)
 titers         0
 filter         -       -    0.0%  (score == utility)
```

**After compaction on `master`:**
```
__level_____count____size___score______in__ingest(sz_cnt)____move(sz_cnt)___write(sz_cnt)____read___r-amp___w-amp
    WAL         1     0 B       -     0 B       -       -       -       -     0 B       -       -       -     0.0
      0         0     0 B    0.00     0 B     0 B       0     0 B       0     0 B       0     0 B       0     0.0
      1         0     0 B    0.00     0 B     0 B       0     0 B       0     0 B       0     0 B       0     0.0
      2         0     0 B    0.00     0 B     0 B       0     0 B       0     0 B       0     0 B       0     0.0
      3         0     0 B    0.00     0 B     0 B       0     0 B       0     0 B       0     0 B       0     0.0
      4         0     0 B    0.00     0 B     0 B       0     0 B       0     0 B       0     0 B       0     0.0
      5         0     0 B    0.00     0 B     0 B       0     0 B       0     0 B       0     0 B       0     0.0
      6      1794   111 G       -     0 B     0 B       0     0 B       0     0 B       0     0 B       0     0.0
  total      1794   111 G       -     0 B     0 B       0     0 B       0     0 B       0     0 B       0     0.0
  flush         0
compact         0     0 B     0 B       0          (size == estimated-debt, score = in-progress-bytes, in = num-in-progress)
  ctype         0       0       0       0       0  (default, delete, elision, move, read)
 memtbl         0     0 B
zmemtbl         0     0 B
   ztbl         0     0 B
 bcache         0     0 B    0.0%  (score == hit-rate)
 tcache         0     0 B    0.0%  (score == hit-rate)
  snaps         0       -       0  (score == earliest seq num)
 titers         0
 filter         -       -    0.0%  (score == utility)
```

**After compaction on `concurrent-manual`:**
```
__level_____count____size___score______in__ingest(sz_cnt)____move(sz_cnt)___write(sz_cnt)____read___r-amp___w-amp
    WAL         1     0 B       -     0 B       -       -       -       -     0 B       -       -       -     0.0
      0         0     0 B    0.00     0 B     0 B       0     0 B       0     0 B       0     0 B       0     0.0
      1         0     0 B    0.00     0 B     0 B       0     0 B       0     0 B       0     0 B       0     0.0
      2         0     0 B    0.00     0 B     0 B       0     0 B       0     0 B       0     0 B       0     0.0
      3         0     0 B    0.00     0 B     0 B       0     0 B       0     0 B       0     0 B       0     0.0
      4         0     0 B    0.00     0 B     0 B       0     0 B       0     0 B       0     0 B       0     0.0
      5         0     0 B    0.00     0 B     0 B       0     0 B       0     0 B       0     0 B       0     0.0
      6      1773   111 G       -     0 B     0 B       0     0 B       0     0 B       0     0 B       0     0.0
  total      1773   111 G       -     0 B     0 B       0     0 B       0     0 B       0     0 B       0     0.0
  flush         0
compact         0     0 B     0 B       0          (size == estimated-debt, score = in-progress-bytes, in = num-in-progress)
  ctype         0       0       0       0       0  (default, delete, elision, move, read)
 memtbl         0     0 B
zmemtbl         0     0 B
   ztbl         0     0 B
 bcache         0     0 B    0.0%  (score == hit-rate)
 tcache         0     0 B    0.0%  (score == hit-rate)
  snaps         0       -       0  (score == earliest seq num)
 titers         0
 filter         -       -    0.0%  (score == utility)
```
</details>

<details>
  <summary>Machine</summary>

<br>

**Machine type**: n1-standard-16
**SSD**: 1000GB pd-ssd
</details>

<details>
  <summary>Resource Utilization</summary>

<br>

![SCR-20220215-tqe](https://user-images.githubusercontent.com/19630228/154184601-dcb50d5c-2fde-4d80-bcfb-70d860c1a44d.png)

![SCR-20220215-tsb](https://user-images.githubusercontent.com/19630228/154184771-b27bdc0b-6ee6-47c5-bb5a-7ea96e64279e.png)

</details>